backend/pyproject.toml
[project]
name = "compen-backend"
version = "0.1.0"
description = "compen backend - collect and visualize pipeline/service events"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.21.0",
    "python-ulid>=3.0.0",
    "pydantic>=2.0.0",
    "openai>=2.31.0",
    "python-dotenv>=1.0.0",
    "PyJWT>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
dev-dependencies = []
